chore: move IdentityTests to Domain tests
